pkp/pkp-lib#9345 allow timeout jobs to be retired by setting failOnTimeout to false and moved context aware query to service provider
